SREP-923: Change dir/filenames & add team name to rules
This change addresses PR feedback to flatten the directory structure
more and to instead incorporate into the filename what the thing is for
(instead of putting it in the directory name). This change will also add
SREP/SREO to the rule names so it is clear where they come from and do
not possibly collide with rule names originating from other teams.

Signed-off-by: Lisa Seelye <lseelye@redhat.com>