Add branding for console login.
When applying changes in k8s an annotation is created w/ the previous state.  This has a limit of about 250k chars.  The login branding lands as secrets and the files are pretty big.  In order to have this be managed successfully we have to break those templates into separate resources, or oc apply will fail.  This means hive SelectorSyncSet also fails with a very large resource.

Therefore I have refactored a bit so that each sub-dir of deploy/ becomes a separate SelectorSyncSet in the Template.

There is one SelectorSyncSet added in the template that is not represented in deploy/.  It is to replace the older single SSS, `osd-managed-cluster-config`.  This means when we deploy this template no manual changes are required to disable the old one.