Do not manage CatalogSourceConfig from Hive
https://jira.coreos.com/browse/SREP-1884

The CatalogSourceConfig that is required to install the operators will be updated by installation of additional operators in the cluster.  Hive will clobber those changes when it reconciles.  Therefore the easiest thing, now that customers can install operators, is to have them also install those two for the logging stack.  We will keep the Namespace and curator ConfigMap creation in place.