🎨: Refactor vercomp and fix minor :bug:
This commit is largely optional, and the byproduct of a 10 minute rabbit
hole while reviewing the upgrade script in detail as part of SDE-170.

If applied, this will force the bin/bash dependency (already present in
the shebang) due to the `<(...)` input redirection.

Technically, this takes advantage of `sort(1)`'s `-v|--version-sort`
feature since it does all the heavy lifting and handles all kinds of
version strings. The interface has not changed at all, just the inner
workings.

Given two versions, `requested` (`1.2.3`) and `default` (`1.2.2`), we
should expect a return value of 1 and when that happens, we know that
when the two versions are sorted among themselves, `default` will come
first and `requested` will come last:

```
1.2.2 <-- default
1.2.3 <-- requested
```

Thus, if we look at the last line of the sort and find that it is the
same as the requested version, we know it's too far ahead of the default
version and should return a 1 to block the operation. If the two
versions are equal, we have a special case where that is allowed, and
thus return 0.

The :bug: fix is this:

When `requested` = `1.2.3-r2` and `default` = `1.2.3-r1`, we can see
that r2 is more recent than r1 and it should be disallowed based on the
previous logic, however, the previous version will return a 0 to allow
the operation. By utilising `sort(1)`, this behaviour, and other oddball
version strings can be caught.

Signed-off-by: Lisa Seelye <lseelye@redhat.com>