Put the DaemonSet back in `kube-system`
...for now, because it doesn't work otherwise.

Per a [conversation in kubernetes slack](https://kubernetes.slack.com/archives/C8EJ01Z46/p1587135386129600?thread_ts=1587127218.128100&cid=C8EJ01Z46)
that may be because the DaemonSet is in the `system-node-critical`
priority class.

I've got other changes I want to make, so I'm reverting this for now
with the goal of reinstating it in a future commit.