efs-csi: use `latest` CSI driver image
In testing it was discovered that aws-efs-csi-driver:release-0.3 was
built using amazon-efs-utils 1.21, which doesn't include a dependency on
openssl, and therefore doesn't support TLS, which is required in order
to mount with access points.

Requesting a new tagged build via [this
issue](https://github.com/kubernetes-sigs/aws-efs-csi-driver/issues/152),
whereupon we'll revert this commit.