Apply extended-dedicated-admins in all staging clusters
Eventually we plan to roll this out to all customer clusters in production but we can't yet.
Until then, roll it out to all of staging.
NOTE this explicitly does not include integration environment because the "or" logic uses the label key for the SSS name and therefore cannot support multiple "or" conditions on the same label.  We have not had asks for integration so the quick update is to just do this for staging.