Add workload monitoring label to openshift-customer-monitoring namespace
Related to https://issues.redhat.com/browse/APPSRE-2537
Also see: https://gitlab.cee.redhat.com/service/app-sre-observability/-/merge_requests/202

When serviceMonitorNamespaceSelector and ruleNamespaceSelector are specified, the selectors are ANDed together.

We need a label that doesn't conflict with the cluster monitoring label. This change is in line with what's in the actual user workload monitoring feature