osd-cluster-ready: restartPolicy: OnFailure
Change the `restartPolicy` for osd-cluster-ready to `OnFailure` so that
it will restart if errors occur. Remove the overrides for `backoffLimit`
and `activeDeadlineSeconds` to let the defaults for those take over.

(Note that this is not really related to the maximum cluster age
observed by the job. The job only fails if an error occurs, not if
health checks are unsuccessful. They tie together only in the sense
that, if the job restarts after one or more failures, it still uses the
inception of the cluster (not the job) to determine the maximum age.)