OSD-6348: Add CuratorJobFailedSRE alert
Peel out a special case of KubeJobFailed for logging curator jobs, which
alert when a Job fails.

The curator job runs on a 1h interval. If one fails, but the next
succeeds, that constitutes a "recovery" which should be considered a
happy path in a k8s context. So this alert only fires if we see two
consecutive failures (over two hours, to account for the cronjob
interval).

OSD-6348