[MT-SRE 400] Stop creating openshift-logging namespace on OSD bootstrap
This PR removes:

1) deploy/bz1905684
2) deploy/osd-logging/unsupported/
3) SSS related to bz1905684 from hack/osd-stage.yaml

The workaround bz1905684(a cronjob) was possibly added because of an [olm
bug](https://github.com/operator-framework/operator-lifecycle-manager/pull/2209 ) that has been fixed in 4.7.
The conjob no longer needs to be present on cluster, and therefore the
openshift-logging namespace does not need to get created on OSD bootstrap.