deploy/OCPBUGS-1341-PodNetworkConnectivityCheck-leak: CronJob pruner
Similar to the pruner in [1], as linked from [2] (the sre- name prefix
avoids collisions if both [1] and this resource set are deployed on
the same cluster).  Limiting to 4.8+ because as Matt points out [3]:

* openshift-network-diagnostics does not exist on 4.6 and
* batch/v1 isn't available until 4.8.

And removing service-cluster to avoid the other issue Matt points out
[3], where we don't want to remove entries like
network-check-source-ip-10-0-211-116-to-network-check-target-service-cluster.

I don't want to bother with containers to build the hack/* stuff, so I
generated that with:

  $ pip3 install --user oyaml
  $ python3 scripts/generate_template.py -t scripts/templates/ -y deploy -d "${PWD}/hack/" -r managed-cluster-config

[1]: https://github.com/openshift/release/pull/32338
[2]: https://issues.redhat.com/browse/OCPBUGS-1341
[3]: https://github.com/openshift/managed-cluster-config/pull/1278#issuecomment-1251936944