deploy/sre-build-test: Set 'ttlSecondsAfterFinished: 86400'
So that failing jobs only trigger KubeJobFailed for a day before they
get cleaned up.  Ideally somebody notices the failures and checks in
to see what's going on.  But if everyone is busy with more important
things, and subsequent CronJob runs have been passing, just clean up
the mess so folks don't have to do that manually.  And if the
underlying issue crops up again later, folks will have another chance
to go investigate the failure mode then.

This supplements the existing self-cleaning functionality that had
been reaping earlier jobs after a successful run, which we've had in
one form or another since f075671565 (Clean up old jobs when
build-test is completed, 2021-10-14, #964).  I'm leaving that logic in
place, because it will help us stop alerting more quickly, without
having to wait out the remainder of the day, if we have a successful
run.

Docs for the property in [1].

I don't want to bother with containers to build the hack/* stuff, so I
generated that with:

  $ pip3 install --user oyaml
  $ python3 scripts/generate_template.py -t scripts/templates/ -y deploy -d "${PWD}/hack/" -r managed-cluster-config

[1]: https://kubernetes.io/docs/concepts/workloads/controllers/job/#ttl-mechanism-for-finished-jobs