Allow CEE to view Addons
This PR is for the CEE team and allows us to view and list child objects of Addons. The change consists of displaying any resource that is a child of the "addons.managed.openshift.io" apiGroup. With this new feature, CEE will be able to easily view and enumerate the various resources associated with Addons, and improve our overall workflow.

This request originated from https://issues.redhat.com/browse/OSD-14426