Modify generate_template.py to skip some files.
Adding the config.yaml with deploymentMode: Policy will skip all files
in the directory.

Added so policy generation files will not be added from their sources.

Also added regex to skip the generated files, so some files can only be
deployed on hypershift-management clusters, but not to the production,
staging or integration hives.