OSD-13913
Removed rosa-console-branding from the generate-policy-config.py script as to not overwrite the policy in the future.
Ran make to generate the hack/* files.
Renamed rosa-console-branding.Policy.yaml by removing the GENERATED part of the name since it's no longer generated.  scripts/generate-policy-config.py was responsible for the original generation of the policy file.