'roleRef' PR check excludes 'namespace'
The 'namespace' property we have added in many roleRef is not a property
in the specification of RoleBinding.  ACM Policy validation fails
because of this.  Therefore we need to remove this invalid and
ineffectual property (it doesn't get persisted in cluster at all).  This
change permits the PR check for roleRef changes to ignore changes to
.roleRef.namespace.