Creates new FileDescriptorLimit alerts
This creates two new alerts to track the NodeFileDescriptor limit based
on what role the instance has. This allows us to continue to handle both
of these as is now, but we can then pull the worker-node one into OCM
Agent Operator in the future to automatically send the alert to
customers.