update infra resize alert (#1784)
* Renames infra resizing rules and extends CPU alerting threshold

This PR renames the infra resizing expression names to be more in line with what they're actually reporting, and renames the alerts triggered for the same reason.  These alerts will not make it to Primary on-call, because they're warnings, but will be used to trigger a single InfraNodeNeedsResizingSRE alert that will roll up to Primary.

This also extends the CPU `for` threshold from 8 hours to 16 hours, to account for increased "daytime" usage that may be causing some flapping alerts.  The longer threshold will ensure excessive CPU utilization continues for more than just the workday before triggering a "resize required" alert.

Finally, this adds an average of the values of the CPU excessive resource consumption, so we receive just one alert, rather than one for every node in the cluster that is currently firing.

Signed-off-by: Chris Collins <collins.christopher@gmail.com>

* on push: make

---------

Signed-off-by: Chris Collins <collins.christopher@gmail.com>
Co-authored-by: clcollins <clcollins@users.noreply.github.com>