update ControlPlaneNodesNeedResizingSRE's warning level alert
This commit adds the recording rules:
* sre:node_control_plane:excessive_consumption_memory
* sre:node_control_plane:excessive_consumption_cpu

which will evaluate to true if the overall memory or CPU utilization of
the control plane nodes is greater than (n-1)/n where n is the number of
control plane nodes. Since we only support 3 control plane nodes this
should always be 2/3. The warning level version of the alert has been
updated to take these two expressions into account, just like what was
done with InfraNodesNeedResizingSRE in #1784.

It also removes the recording rule:
* sre:node_roles:node_num_cpu

as it is no longer used anywhere.

Signed-off-by: Michael Shen <mshen@redhat.com>