Create a specific MHC for management clusters
This MHC will be specifically used to recover request-serving nodes on
the Hosted Control Plane Management Clusters. It's needed because the
HyperShift Operator currently deletes underlying nodes when a
hostedcluster is uninstalled and specifying no .spec.unhealthyConditions
allows a MHC to detect this without overlapping with other unhealthy
node conditions that can be caused by widespread issues such as AZ
outages.

Signed-off-by: Michael Shen <mshen@redhat.com>