Add `iam:GetPolicyVersion` to the OCM admin role
For the auto admin flow, the backend creates the operator roles
for the user.
In the case of shared VPC, we need to validate existing ingress operator
policies to make sure not overriding an exisiting shared VPC role ARN.
Therefore, we need to fetch the policy document via API call `GetPolicyVersion`.

Related: OCM-4267