OCM-2303 | [HCP] Disable ingress certificate in ACM policy if configmap parameter is true
Ingress in HCP are configured by an ACM policy and configuration are passed in a configmap written by OCM.

In local environemnt, certificates are not generated, so specifying them in the ingress config make the operator
degraded, as the relevant secret does not exist. To fix this, we added a new parameter in the CM to skip certificates.

With this change, the policy will read the parameter and exclude any certificate related config/object if it is true.
If it is false or not present, we will behave as before (so reference the certificates). So backward compatibility
is guaranteed. No impact expected in standard integration/staging/production clusters.

Tested the changes in an integration Service Cluster.