remove unneeded kms:DescribeKey permission from 4.14+
The _Creating a ROSA cluster with STS using customizations_ doc [0]
instructs customers to grant the necessary permissions (including
kms:DescribeKey) in the key policy for the key they want to use. Since
this permission is already granted in the key policy, we don't need it
granted in the control plane policy as well.

This also allows us to limit the privileges we have, as we don't get
access to "*" (all kms keys) and don't require the customer to do any
additional tagging of their keys.

AWS-managed keys will not be affected by this, as all principals in the
account that are authorized to use EC2 have permission to use
AWS-managed keys.

[0]
https://docs.openshift.com/rosa/rosa_install_access_delete_clusters/rosa-sts-creating-a-cluster-with-customizations.html#rosa-sts-creating-cluster-customizations-cli_rosa-sts-creating-a-cluster-with-customizations