Restor MGO's NS, but for fedramp only
There are parts of MCC that assume the NS is present. In commercial this
works perfectly fine, because the Release Bundle sets up MGO. In
FedRamp MGO does not exist, so the (superflous in FR) configs that
depend on the NS cause SSSes to error out.

The more complex solution would be to make sure those unneeded MGO
NS-dependant configs do not get applied in FR. The easier solution is to
just make sure the namespace exists in FR, so nothing errors out. Hence
this change.

This is a followup fix for #2157