OCPBUGS-55803: Use console-provided ROSA branding
Removes the `rosa-brand-logo.svg` custom logo in favor of setting the `brand` to `ROSA`.

This fixes an issue with the contrast of the ROSA logo on OpenShift 4.19+, where the masthead can change background colours from light to dark.

Instead of using the [new custom logo API](https://github.com/openshift/enhancements/pull/1753), which would contain new schema that is not recognized by older versions of OpenShift, changing the value of `brand` achieves the same effect (i.e., the title and logo are changed).

Any future updates that are needed to the ROSA logo can be made in the openshift/console repo.