SREP-400: Refactor CAPI annotator job This is a refacftor of the CAPI annotator job that we run on the fleet to ensure HCP clusters work with and without the ability to tag ENIs on `RunInstances` calls.
The following changes have been made, with inspiration from the work done in https://github.com/openshift/managed-cluster-config/pull/2353

* Only pin the image for clusters that are in the affected range
* Unpin clusters once they upgrade out of hte affected range
* Consolidate the logic + scripts into a single ConfigMap. Changes to the scripts need to be copied to the `04-ConfigMap.yaml`.

Given we have north of 500 clusters in the affected Z stream ranges, we will need to run this job for some time. By adding
the ability to remove the pin when not needed, we ensure new CAPI changes are applied to the fleet.