Move brand switch to 4.15 to account for race condition during upgrade
During testing for SREP-901 we found that even though the console CO is
reporting 4.14.z and in OCM the cluster is still listed as 4.13.z, the
4.14 syncsets get applied. In this specific case with the console
config and the custom brand image, it removes the image configmap
without issue but then it cannot apply the config changes because the CR
rejects the `brand: rosa` change. This prevents the json patch from
removing the additional brand config, which leads to a degraded console
operator, which leads to the upgrade stalling. This means there's a good
chance we will require SRE intervention on every 4.13 -> 4.14 upgrade.

This commit pushes the configuration to deploy at the 4.14 -> 4.15
upgrade, so that we can _ensure_ that the CRD changes for the console
are present before applying them and removing the old configuration,
preventing the cluster from becoming stuck.